+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking
 
 * the definition of `ampsvdtol` threshold for SVD methods has been changed. Now it corresponds to the threshold for the density matrix (i.e., square of the previous definition).
+* `verbosity` has been moved from `ECinfo` to `Options.print.time`.
 
 ### Changed